Refactor fingerprint handling
This commit refactors the fingerprint handling logic in `MainActivity.java`.
It introduces a new `lastFingerprintTemplateB64` field to store the latest fingerprint template.
A new private method `sendLatestFingerprintToWeb` is added to consolidate the logic for sending both the fingerprint image and template to the WebView.
This new method is now called from both `onImageReceived` and `onCapture` callbacks of the `FingerprintSensorListener`.
Previously, the JavaScript call to `window.AjoibotFinger` was made directly within the `onCapture` callback.